% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_categories.R
\name{tg_categories}
\alias{tg_categories}
\title{Category list}
\usage{
tg_categories(lang = NULL)
}
\arguments{
\item{lang}{Response language}
}
\value{
tibble with categories
}
\description{
List of 'TGStat' channel categories
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/database/categories.html}{TGstat API Documentation of metrod database/categories}
}
