% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_channel.R
\name{tg_channel}
\alias{tg_channel}
\title{Get chennel info}
\usage{
tg_channel(channel_id)
}
\arguments{
\item{channel_id}{hannel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}
}
\value{
tibble with channel metadata
}
\description{
Get general information about the channel - link to the channel, name, description, avatar, number of subscribers at the moment.
}
\examples{
\dontrun{
channel <- tg_channel(channel_id = "R4marketing")
}
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/channels/subscribers.html}{TGstat API Documentation of metrod channels/get}
}
