% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_channel_forwards.R
\name{tg_channel_forwards}
\alias{tg_channel_forwards}
\title{Getting a list of reposts from a channel}
\usage{
tg_channel_forwards(
  channel_id = tg_get_channel_id(),
  start_date = Sys.Date() - 15,
  end_date = Sys.Date()
)
}
\arguments{
\item{channel_id}{Channel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}

\item{start_date}{Date forwards from}

\item{end_date}{Date forwards to}
}
\value{
tibble with forwards
}
\description{
Allows you to get a list of reposts of publications from a channel to other channels.
}
\examples{
\dontrun{
forwards <- tg_channel_forwards(
    channel_id = 'R4marketing',
    start_date = '2021-01-01',
    end_date   = '2021-09-30'
)
}
}
