% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_channel_stat.R
\name{tg_channel_stat}
\alias{tg_channel_stat}
\title{Get channel stat}
\usage{
tg_channel_stat(channel_id)
}
\arguments{
\item{channel_id}{Channel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}
}
\value{
tibble with channel stat
}
\description{
Get channel stat
}
