% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_countries.R
\name{tg_countries}
\alias{tg_countries}
\title{List of countries}
\usage{
tg_countries(lang = NULL)
}
\arguments{
\item{lang}{Response language}
}
\value{
tibble with countries
}
\description{
List of countries 'TGStat'
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/database/countries.html}{TGStat API Documentation of metrod database/countries}
}
