% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_languages.R
\name{tg_languages}
\alias{tg_languages}
\title{List of languages}
\usage{
tg_languages(lang = NULL)
}
\arguments{
\item{lang}{Response language}
}
\value{
tibble
}
\description{
List of available languages for 'TGStat' channels
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/database/languages.html}{TGStat API Documentation of metrod database/languages}
}
