% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_post.R
\name{tg_post}
\alias{tg_post}
\title{Retrieving publication data}
\usage{
tg_post(post_id)
}
\arguments{
\item{post_id}{Post ID (t.me/username/123, t.me/c/1256804429/1230 or post ID in TGStat)}
}
\value{
tibble with post data
}
\description{
Retrieving publication data
}
\details{
Get information and publications in Telegram - number of views, publication date, content, ...
}
\examples{
\dontrun{
post <- tg_post(
    post_id = 'https://t.me/R4marketing/887'
)
}
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/posts/get.html}{TGStat API Documentation of metrod posts/get}
}
