% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{tg_set_api_quote_alert_rate}
\alias{tg_set_api_quote_alert_rate}
\title{Set API limit alert rate}
\usage{
tg_set_api_quote_alert_rate(api_quote_alert_rate)
}
\arguments{
\item{api_quote_alert_rate}{Max reach of API limit to alert}
}
\value{
using for side effect, no returm value
}
\description{
Set API limit alert rate
}
