% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{tg_set_channel_id}
\alias{tg_set_channel_id}
\title{Set session default channel id}
\usage{
tg_set_channel_id(channel_id)
}
\arguments{
\item{channel_id}{Channel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}
}
\value{
Using for side effect, no return data
}
\description{
Set session default channel id
}
\examples{
\dontrun{
tg_set_channel_id('R4marketing')
stat <- tg_channel_stat()
}
}
