% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{tg_set_check_api_quote}
\alias{tg_set_check_api_quote}
\title{Disable or enable API limit alert}
\usage{
tg_set_check_api_quote(check_api_quote)
}
\arguments{
\item{check_api_quote}{Logical, disable (or enable) API limit alerts}
}
\value{
using for side effect, no return value
}
\description{
Disable or enable API limit alert
}
