% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{tg_set_interval}
\alias{tg_set_interval}
\title{Set time interval in seconds between tries of HTTP queries}
\usage{
tg_set_interval(interval)
}
\arguments{
\item{interval}{delay between retries}
}
\value{
using for side effect, no returm value
}
\description{
Set time interval in seconds between tries of HTTP queries
}
