% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{tg_set_max_tries}
\alias{tg_set_max_tries}
\title{Set max tries of HTTP queries}
\usage{
tg_set_max_tries(max_tries)
}
\arguments{
\item{max_tries}{integer, maximum number of attempts}
}
\value{
using for side effect, no returm value
}
\description{
Set max tries of HTTP queries
}
