% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgstat_package.R
\docType{package}
\name{rtgstat-package}
\alias{rtgstat}
\alias{rtgstat-package}
\title{rtgstat: Client for 'TGStat API'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Allow function for using 'TGStat Stat API' and 'TGStat Search API', for more details see <https://api.tgstat.ru/docs/ru/start/intro.html>. 'TGStat' provide telegram channel analytics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://selesnow.github.io/rtgstat/}
  \item Report bugs at \url{https://github.com/selesnow/rtgstat/issues}
}

}
\author{
\strong{Maintainer}: Alexey Seleznev \email{selesnow@gmail.com} (\href{https://orcid.org/0000-0003-0410-7385}{ORCID})

}
