% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_auth.R
\name{tg_auth}
\alias{tg_auth}
\title{Set API Token of 'TgStat'}
\usage{
tg_auth(token)
}
\arguments{
\item{token}{Your API token.}
}
\value{
Use only for set token. No return value.
}
\description{
Set API Token of 'TgStat'
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/start/token.html}{TGStat API Documentation of Authorization}
}
