% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_channel.R
\name{tg_channel}
\alias{tg_channel}
\title{Get chennel info}
\usage{
tg_channel(channel_id)
}
\arguments{
\item{channel_id}{hannel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}
}
\value{
tibble with channel metadata
}
\description{
Get chennel info
}
