% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_channel_stat.R
\name{tg_channel_stat}
\alias{tg_channel_stat}
\title{Get channel stat}
\usage{
tg_channel_stat(channel_id = tg_get_channel_id())
}
\arguments{
\item{channel_id}{Channel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}
}
\value{
tibble with channel stat
}
\description{
The method allows you to obtain basic statistics - the number of participants, the average coverage of the publication, the percentage of engagement of subscribers (ERR), the total daily coverage, the citation index (CI)
}
\examples{
\dontrun{
channel_stat <- tg_channel_stat(channel_id = "R4marketing")
}
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/channels/stat.html}{TGstat API Documentation of metrod channels/stat}
}
