% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tg_channel_subscribers.R
\name{tg_channel_subscribers}
\alias{tg_channel_subscribers}
\title{Get channel subscribers nunmber by day}
\usage{
tg_channel_subscribers(
  channel_id,
  start_date = Sys.Date() - 15,
  end_date = Sys.Date(),
  group = c("day", "hour", "week", "month")
)
}
\arguments{
\item{channel_id}{Channel ID (@username, t.me/username, t.me/joinchat/AAAAABbbbbcccc ... or channel ID in 'TGStat')}

\item{start_date}{Start date of report period}

\item{end_date}{End date of report period}

\item{group}{Time group: hour, day, week, month}
}
\value{
tibble with subscribers stat
}
\description{
The method allows you to get the number of channel subscribers in dynamics by hours, days, weeks, months.
}
\references{
See also \href{https://api.tgstat.ru/docs/ru/channels/subscribers.html}{TGstat API Documentation of metrod channels/subscribers}
}
